+++ rfm69-async/Cargo.toml
@@ -13,9 +13,14 @@ categories = [
     "embedded-hal-async",
 ]
 
+[features]
+embassy = ["dep:embassy-time"]
+
 [dependencies]
 log = "0.4"
 
 embedded-hal-1 = { package = "embedded-hal", version = "=1.0.0-alpha.10" }
 embedded-hal-async = "0.2.0-alpha.1"
 heapless = "0.7.16"
+
+embassy-time = { version = "0.1.0", features = ["defmt", "defmt-timestamp-uptime", "unstable-traits", "nightly"], optional = true }
